fix: separate resolvers that depend on ogm from other resolvers
