+++ libs/backend/infra/adapter/neo4j/src/resolver/index.ts
@@ -1,3 +1,2 @@
-export * from './element'
-export * from './page'
+export * from './pure-resolver'
 export * from './resolvers'
