+++ libs/backend/infra/adapter/neo4j/src/resolver/ogm-resolver/domain/domain.resolver.ts
@@ -1,13 +1,7 @@
 import type { IResolvers } from '@graphql-tools/utils'
-import { domainConfig } from './field/domain-config'
-import { projectDomain } from './field/project-domain'
 import { createDomains, deleteDomains, updateDomains } from './mutation'
 
 export const domainResolver: IResolvers = {
-  Domain: {
-    domainConfig,
-    projectDomain,
-  },
   Mutation: {
     createDomains,
     deleteDomains,
