+++ libs/backend/infra/adapter/neo4j/src/resolver/ogm-resolver/domain/mutation/create-domains.ts
@@ -2,8 +2,8 @@ import type { GraphQLRequestContext } from '@codelab/backend/abstract/types'
 import { vercelApis } from '@codelab/backend/infra/adapter/vercel'
 import type { OGM_TYPES } from '@codelab/shared/abstract/codegen'
 import type { IFieldResolver } from '@graphql-tools/utils'
-import { Repository } from '../../../infra'
-import { domainSelection } from '../../../selectionSet'
+import { Repository } from '../../../../infra'
+import { domainSelection } from '../../../../selectionSet'
 import { domainExistsError } from '../validation/domain.error'
 
 export const createDomains: IFieldResolver<
