+++ libs/backend/infra/adapter/neo4j/src/resolver/ogm-resolver/ogm-resolvers.ts
@@ -1,0 +1,13 @@
+import { mergeResolvers } from '@graphql-tools/merge'
+import type { IResolvers } from '@graphql-tools/utils'
+import { atomResolver } from './atom'
+import { domainResolver } from './domain'
+import { elementResolver } from './element'
+import { tagResolver } from './tag'
+
+export const ogmResolvers: IResolvers = mergeResolvers([
+  atomResolver,
+  domainResolver,
+  elementResolver,
+  tagResolver,
+])
