+++ libs/backend/infra/adapter/neo4j/src/resolver/ogm-resolver/tag/field/tag-descendants.ts
@@ -1,9 +1,9 @@
 import type { Tag } from '@codelab/shared/abstract/codegen'
 import type { IFieldResolver } from '@graphql-tools/utils'
 import type { Node } from 'neo4j-driver'
-import { tagDescendants } from '../../../cypher'
-import { Repository, withReadTransaction } from '../../../infra'
-import { tagSelectionSet } from '../../../selectionSet'
+import { tagDescendants } from '../../../../cypher'
+import { Repository, withReadTransaction } from '../../../../infra'
+import { tagSelectionSet } from '../../../../selectionSet'
 
 export const descendants: IFieldResolver<Tag, unknown> = (parent) =>
   withReadTransaction(async (txn) => {
