+++ libs/backend/infra/adapter/neo4j/src/resolver/pure-resolver/domain/domain.resolver.ts
@@ -1,0 +1,10 @@
+import type { IResolvers } from '@graphql-tools/utils'
+import { domainConfig } from './field/domain-config'
+import { projectDomain } from './field/project-domain'
+
+export const domainResolver: IResolvers = {
+  Domain: {
+    domainConfig,
+    projectDomain,
+  },
+}
