+++ libs/backend/infra/adapter/neo4j/src/resolver/pure-resolver/domain/index.ts
@@ -1,0 +1,1 @@
+export * from './domain.resolver'
