+++ libs/backend/infra/adapter/neo4j/src/resolver/pure-resolver/element/element.resolver.ts
@@ -1,0 +1,10 @@
+import type { IResolvers } from '@graphql-tools/utils'
+import { renderType } from './field/render-type'
+
+export const elementResolver: IResolvers = {
+  Element: {
+    renderType,
+  },
+  Mutation: {},
+  Query: {},
+}
