+++ libs/backend/infra/adapter/neo4j/src/resolver/pure-resolver/element/index.ts
@@ -1,0 +1,1 @@
+export * from './element.resolver'
