+++ libs/backend/infra/adapter/neo4j/src/resolver/pure-resolver/index.ts
@@ -1,0 +1,1 @@
+export * from './pure-resolvers'
