+++ libs/backend/infra/adapter/neo4j/src/resolver/pure-resolver/type/query/base-types.ts
@@ -2,8 +2,8 @@ import type { QueryBaseTypesArgs } from '@codelab/shared/abstract/codegen'
 import type { IFieldResolver } from '@graphql-tools/utils'
 import type { GraphQLRequestContext } from 'graphql-request/dist/types'
 import { int } from 'neo4j-driver'
-import { getBaseTypes } from '../../../cypher'
-import { withReadTransaction } from '../../../infra'
+import { getBaseTypes } from '../../../../cypher'
+import { withReadTransaction } from '../../../../infra'
 
 export const baseTypes: IFieldResolver<
   GraphQLRequestContext,
