+++ libs/backend/infra/adapter/neo4j/src/resolver/pure-resolver/type/query/index.ts
