+++ libs/backend/infra/adapter/neo4j/src/resolver/pure-resolver/type/type.resolver.ts
@@ -1,8 +1,9 @@
 import type { IResolvers } from '@graphql-tools/utils'
+import { baseTypes } from './query'
 
 export const typeResolver: IResolvers = {
   Mutation: {},
   Query: {
-    baseTypes: withReadTransactionResolver(baseTypes),
+    baseTypes: baseTypes,
   },
 }
