+++ libs/backend/infra/adapter/neo4j/src/resolver/resolvers.ts
@@ -1,16 +1,9 @@
 import { mergeResolvers } from '@graphql-tools/merge'
 import type { IResolvers } from '@graphql-tools/utils'
-import { appResolver } from './app'
-import { atomResolver } from './atom'
-import { domainResolver } from './domain'
-import { elementResolver } from './element'
-import { pageResolver } from './page'
+import { ogmResolvers } from './ogm-resolver/ogm-resolvers'
+import { pureResolvers } from './pure-resolver/pure-resolvers'
 
 export const resolvers: IResolvers = mergeResolvers([
-  appResolver,
-  pageResolver,
-  elementResolver,
-  atomResolver,
-  pageResolver,
-  domainResolver,
+  pureResolvers,
+  ogmResolvers,
 ])
